Changes to support reability and clarity
Added a link to the Matrix lobby to make it easier to find support but the chances are they already know where it is from the torrent release. However, there will likely be people who slip through the net.

Corrected capitalization (OCD moment).

Tidied up the Supported documentation and added an introduction to make it clearer.

Reformatted the Unsupported Distributions adding verbosity where required and reducing negative content.

Added an introduction to Hardware Support and expanded the specifications to be more specific and in a format easier to read.

Switched from "-" to Markdown bulletpoints.

Changed the "settings.sh" command area, by adding an introduction and then changing the bash settings command to look more like the `-h` flag available in the shell.

Took out some unnecessary bullet points. 

Used code takes in the Modding Games with Dwarfs area, this needs expanding to make it clear how mods are added.

Reformatted the Additional Information area to increase readability and highlighted the GUI Library. I suspect this should have it's own area.